db: close four small schema defects
Four unrelated audit findings, none big enough for its own revision and all pure corrections to
columns already in use.

`api_keys.created_by_account_id` declared no `ON DELETE` action, so it defaulted to NO ACTION
while `owner_account_id` on the same table was SET NULL. Deleting an account that had ever minted
a key failed outright -- the opposite of what every other `*_by_account_id` audit column in this
schema does, and at odds with the column's own docstring, which records the creator "when known".

`starboards.colour` was BIGINT for a value its own CHECK constraint bounds to 0..16777215, a
24-bit RGB triple. INTEGER holds it with room to spare.

`accounts.created_at`, `builds.submission_time` and `builds.edited_time` were the only three
nullable timestamp columns in the schema carrying a `now()` server default. Nothing can
legitimately read NULL from a column that defaults to the current time; they were nullable
because the reflected baseline was, and the models had been repeating it ever since. The
migration backfills any straggling NULLs before applying the constraint.

+++ squid/starboard/infrastructure/models.py
@@ -68,7 +68,7 @@ class Starboard(Base, kw_only=True):
     link_edits: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("true"), default=True)
     link_deletes: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("true"), default=True)
     display_emoji: Mapped[str] = mapped_column(Text, nullable=False, server_default=text("'⭐'"), default="⭐")
-    colour: Mapped[int] = mapped_column(BigInteger, nullable=False, server_default=text("4415105"), default=0x435E81)
+    colour: Mapped[int] = mapped_column(Integer, nullable=False, server_default=text("4415105"), default=0x435E81)
     jump_to_message: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("true"), default=True)
     attachments_list: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("true"), default=True)
     replied_to: Mapped[bool] = mapped_column(Boolean, nullable=False, server_default=text("true"), default=True)
